requirements.txt
connexion[swagger-ui]==2.7.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
python-dotenv==0.14.0
flask-session~=0.3.2
requests>=2,<3
msal>=0.6.1,<2
flask-cors == 3.0.9
pyjwt[crypto]==1.7.1